# internal/server/openapi.yaml
openapi: 3.1.0
info:
  title: Varkiv HTTP API
  summary: Stable, versioned contract for the Varkiv library and device ecosystem
  version: 0.1.0-preview.1
  description: |
    Versioned HTTP API for a self-hosted personal ROM catalog, frontend imports
    and exports, scoped device pairing, multi-file save streams, and automatic
    sync sessions. Every API response includes `X-Request-ID`,
    `X-Varkiv-API-Version`, and `Cache-Control: no-store`.

    Unless an operation explicitly declares `security: []`, deployments with an
    owner token require `Authorization: Bearer <token>`. Device tokens are opaque,
    independently revocable credentials whose effective scopes are enforced by
    the server. Signed content/session URLs are short-lived capabilities rather
    than bearer tokens.
  license:
    name: Apache License 2.0
    identifier: Apache-2.0
...
servers:
  - url: /api/v1
    description: Same-origin stable API entry point
security:
  - bearerAuth: []
...
paths:
...
  /series:
    get:
      operationId: getSeries
      tags: [Series]
      summary: List cross-platform series with member Games
      parameters:
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Offset'}
        - {$ref: '#/components/parameters/Locale'}
        - {$ref: '#/components/parameters/Projection'}
        - {name: q, in: query, schema: {type: string}, description: "Search series titles, descriptions, member titles, platforms, and relation types."}
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '200': {description: Paginated full or summary series, content: {application/json: {schema: {oneOf: [{$ref: '#/components/schemas/SeriesCollection'}, {$ref: '#/components/schemas/SeriesSummaryCollection'}]}}}}
    post:
      operationId: postSeries
      tags: [Series]
      summary: Create a cross-platform series and its optional complete member set atomically
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SeriesMutation'}
      responses:
        '201': {description: Created series, content: {application/json: {schema: {$ref: '#/components/schemas/Series'}}}}
        '400': {$ref: '#/components/responses/Error'}
...
  /web-emulation/saves/{token}:
    parameters:
      - {name: token, in: path, required: true, schema: {type: string}}
    get:
      operationId: getWebEmulationSavesByToken
      tags: [Saves]
      summary: Restore the current save for a signed browser-play session
      security: []
      responses:
        '200':
          description: Current core-isolated battery save
          headers:
            X-Varkiv-Revision: {schema: {type: string}, description: Revision to send as the base of the next upload}
          content: {application/octet-stream: {schema: {type: string, format: binary}}}
        '204': {description: The stream has no save revision yet}
        '401': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
    post:
      operationId: postWebEmulationSavesByToken
      tags: [Saves]
      summary: Add an automatic core-isolated save revision
      description: The signed play capability selects Edition, exact EmulatorJS core, driver, and SaveStream. A stale base is preserved as a conflict instead of overwriting history.
      security: []
      parameters:
        - {name: X-Varkiv-Base-Revision, in: header, schema: {type: string}}
      requestBody:
        required: true
        content: {application/octet-stream: {schema: {type: string, format: binary, maxLength: 33554432}}}
      responses:
        '201': {description: Immutable save revision created or deduplicated, content: {application/json: {schema: {$ref: '#/components/schemas/SavePushResult'}}}}
        '401': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '415': {$ref: '#/components/responses/Error'}
...
  /devices:
    get:
      operationId: getDevices
      tags: [Devices]
      summary: List registered devices
      parameters: [{$ref: '#/components/parameters/Limit'}, {$ref: '#/components/parameters/Offset'}]
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '200': {description: Paginated devices, content: {application/json: {schema: {$ref: '#/components/schemas/DeviceCollection'}}}}
    post:
      operationId: postDevices
      tags: [Devices]
      summary: Register a device
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/DeviceMutation'}}}}
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '201': {description: Registered device, content: {application/json: {schema: {$ref: '#/components/schemas/Device'}}}}
  /devices/{id}:
    parameters: [{$ref: '#/components/parameters/ID'}]
    get:
      operationId: getDevicesById
      tags: [Devices]
      summary: Get a registered device
      responses:
        '200': {description: Device, content: {application/json: {schema: {$ref: '#/components/schemas/Device'}}}}
        '404': {$ref: '#/components/responses/Error'}
    put:
      operationId: putDevicesById
      tags: [Devices]
      summary: Replace editable device metadata
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/DeviceMutation'}}}}
      responses:
        '200': {description: Updated device, content: {application/json: {schema: {$ref: '#/components/schemas/Device'}}}}
        '404': {$ref: '#/components/responses/Error'}
    delete:
      operationId: deleteDevicesById
      tags: [Devices]
      summary: Delete a device that has no save revisions
      responses:
        '204': {description: Deleted}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
...
  /saves:
    get:
      operationId: getSaves
      tags: [Saves]
      summary: List immutable save revisions
      parameters:
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Offset'}
        - {name: edition_id, in: query, schema: {type: string, format: uuid}}
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '200': {description: Paginated save revisions, content: {application/json: {schema: {$ref: '#/components/schemas/SaveRevisionCollection'}}}}
...
  /saves/{id}:
    get:
      operationId: getSavesById
      tags: [Saves]
      summary: Get immutable save revision metadata
      parameters: [{$ref: '#/components/parameters/ID'}]
      responses:
        '200': {description: Save revision metadata, content: {application/json: {schema: {$ref: '#/components/schemas/SaveRevision'}}}}
        '404': {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Owner token or a server-issued device token; device scopes are enforced per route.
  headers:
    RequestID:
      description: Caller-supplied valid request identity or a server-generated identity.
      schema: {type: string, minLength: 1, maxLength: 128, pattern: '^[A-Za-z0-9._-]+$'}
    APIVersion:
      description: Stable API path version used for this response.
      schema: {type: string, const: v1}
  parameters:
    ID: {name: id, in: path, required: true, schema: {type: string}, description: Opaque resource identifier; clients must not infer its shape or ordering.}
    Locale: {name: locale, in: query, schema: {type: string, enum: [zh-CN, zh-TW, ja, en]}, description: Preferred localized title. The resource default remains the fallback.}
    Projection: {name: projection, in: query, schema: {type: string, enum: [full, summary], default: full}, description: "Use summary for list metadata, aggregate artifact state, and an opaque cover reference; fetch the detail resource before editing or launching."}
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 200, default: 100}, description: Maximum number of records returned in this page.}
    Offset: {name: offset, in: query, schema: {type: integer, minimum: 0, default: 0}, description: Zero-based record offset in the endpoint's documented stable ordering.}
...
  responses:
...
    Error:
      description: Structured API error
      headers:
        X-Request-ID: {$ref: '#/components/headers/RequestID'}
        X-Varkiv-API-Version: {$ref: '#/components/headers/APIVersion'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
          example:
            error:
              code: invalid_argument
              message: request payload is invalid
              request_id: 40e39da4b75c41c28cf40b9f8de46c48
  schemas:
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code: {type: string, description: Stable machine-readable error code used for client branching.}
            message: {type: string, description: Human-readable diagnostic; clients must not parse this value.}
            request_id: {type: string, description: Correlation identity also returned in the X-Request-ID header.}
    Pagination:
      type: object
      additionalProperties: false
      description: Offset page metadata. Repeat the same filters and ordering while advancing offset by the number of returned records.
      required: [limit, offset, total]
      properties:
        limit: {type: integer, minimum: 1, maximum: 200}
        offset: {type: integer, minimum: 0}
        total: {type: integer, minimum: 0}
...
    DeviceMutation:
      type: object
      additionalProperties: false
      required: [name, os_family]
      properties:
        id: {type: string}
        name: {type: string, minLength: 1}
        device_profile_id: {type: string}
        os_family: {type: string, minLength: 1}
        distribution: {type: string}
        architecture: {type: string}
        agent_version: {type: string}
        status: {type: string, enum: [active, offline]}
        capabilities: {type: object, additionalProperties: {type: boolean}}
...
    SavePushResult:
      type: object
      additionalProperties: false
      required: [revision, created, conflict]
      properties:
        revision: {$ref: '#/components/schemas/SaveRevision'}
        created: {type: boolean}
        conflict: {type: boolean}
...
    Artifact:
      type: object
      required: [id, edition_id, path, storage_kind, role, size, missing]
      properties:
        id: {type: string, format: uuid}
        edition_id: {type: string, format: uuid}
        path: {type: string}
        storage_kind: {type: string, enum: [library, managed]}
        source_path: {type: string}
        original_name: {type: string}
        role: {type: string, enum: [rom, disc, executable, patch, dlc, update, other]}
        disc_index: {type: integer, minimum: 0, maximum: 64}
        size: {type: integer, format: int64, minimum: 0}
        sha256: {type: string}
        missing: {type: boolean}
    Device:
      type: object
      required: [id, name, os_family, status, capabilities, created_at, updated_at, last_seen_at]
      properties:
        id: {type: string}
        name: {type: string}
        device_profile_id: {type: string}
        os_family: {type: string}
        distribution: {type: string}
        architecture: {type: string}
        agent_version: {type: string}
        status: {type: string, enum: [active, offline, revoked]}
        capabilities: {type: object, additionalProperties: {type: boolean}}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
        last_seen_at: {type: string, format: date-time}
        revoked_at: {type: string, format: date-time}
...
    SaveFile:
      type: object
      required: [id, revision_id, logical_path, checksum, size]
      properties:
        id: {type: string}
        revision_id: {type: string}
        logical_path: {type: string, description: Portable relative path; server blob paths are never exposed.}
        checksum: {type: string, pattern: '^[a-f0-9]{64}$'}
        size: {type: integer, format: int64, minimum: 0}
        mtime_ns: {type: integer, format: int64}
        mode: {type: integer, format: int64}
    SaveRevision:
      type: object
      required: [id, stream_id, content_hash, total_size, file_count, status, files, device_id, created_at]
      properties:
        id: {type: string}
        stream_id: {type: string}
        device_id: {type: string}
        parent_revision_id: {type: string}
        content_hash: {type: string, pattern: '^[a-f0-9]{64}$'}
        total_size: {type: integer, format: int64, minimum: 0}
        file_count: {type: integer, minimum: 1, maximum: 4096}
        status: {type: string, enum: [current, superseded, conflict, quarantined]}
        files: {type: array, minItems: 1, items: {$ref: '#/components/schemas/SaveFile'}}
        created_at: {type: string, format: date-time}
...
    MediaAsset:
      type: object
      required: [id, kind, storage_kind, path, original_name, mime_type, size, sha256, source_type, content_status, created_at]
      properties:
        id: {type: string, format: uuid}
        game_id: {type: string, format: uuid}
        edition_id: {type: string, format: uuid}
        kind: {type: string}
        storage_kind: {type: string, enum: [library, managed]}
        path: {type: string}
        source_path: {type: string}
        original_name: {type: string}
        mime_type: {type: string}
        size: {type: integer, format: int64}
        sha256: {type: string}
        locale: {type: string}
        source_type: {type: string, enum: [upload, frontend-import, provider]}
        sort_order: {type: integer}
        content_status: {type: string, enum: [unverified, available, missing, changed, unsafe]}
        content_checked_at:
          description: RFC 3339 time of the last explicit check, or an empty string when the relation has not been checked yet.
          anyOf:
            - {type: string, format: date-time}
            - {type: string, const: ''}
        created_at: {type: string, format: date-time}
...
    Edition:
      type: object
      required: [id, game_id, default_title, edition_type, artifacts]
      properties:
        id: {type: string, format: uuid}
        game_id: {type: string, format: uuid}
        default_title: {type: string}
        display_title: {type: string}
        titles: {type: object, additionalProperties: {type: string}}
        edition_type: {type: string, enum: [original, translation, hack, revision, homebrew, other]}
        version: {type: string}
        author: {type: string}
        languages: {type: array, items: {type: string}}
        artifacts: {type: array, items: {$ref: '#/components/schemas/Artifact'}}
        media: {type: array, items: {$ref: '#/components/schemas/MediaAsset'}}
    Game:
      type: object
      required: [id, platform, default_title, editions]
      properties:
        id: {type: string, format: uuid}
        platform: {type: string}
        default_title: {type: string}
        display_title: {type: string}
        titles: {type: object, additionalProperties: {type: string}}
        primary_edition_id: {type: string, format: uuid}
        editions: {type: array, items: {$ref: '#/components/schemas/Edition'}}
        media: {type: array, items: {$ref: '#/components/schemas/MediaAsset'}}
    ArtifactStats:
      type: object
      additionalProperties: false
      required: [total, missing, hashed, usable, managed]
      properties:
        total: {type: integer, minimum: 0}
        missing: {type: integer, minimum: 0}
        hashed: {type: integer, minimum: 0}
        usable: {type: integer, minimum: 0, description: Non-missing artifacts with a content hash.}
        managed: {type: integer, minimum: 0}
    MediaSummary:
      type: object
      additionalProperties: false
      required: [id, kind, mime_type]
      description: Opaque thumbnail reference without storage paths, hashes, or source metadata.
      properties:
        id: {type: string, format: uuid}
        kind: {type: string}
        mime_type: {type: string}
    EditionSummary:
      type: object
      additionalProperties: false
      required: [id, game_id, default_title, display_title, edition_type, languages, save_namespace, sort_order, titles, artifact_stats]
      properties:
        id: {type: string, format: uuid}
        game_id: {type: string, format: uuid}
        default_title: {type: string}
        display_title: {type: string}
        edition_type: {type: string, enum: [original, translation, hack, revision, homebrew, other]}
        version: {type: string}
        languages: {type: array, items: {type: string}}
        author: {type: string}
        save_namespace: {type: string}
        serial: {type: string}
        product_code: {type: string}
        title_id: {type: string}
        sort_order: {type: integer}
        titles: {type: object, additionalProperties: {type: string}}
        artifact_stats: {$ref: '#/components/schemas/ArtifactStats'}
    GameSummary:
      type: object
      additionalProperties: false
      required: [id, default_title, display_title, platform, titles, editions, created_at, updated_at]
      properties:
        id: {type: string, format: uuid}
        default_title: {type: string}
        display_title: {type: string}
        platform: {type: string}
        primary_edition_id: {type: string, format: uuid}
        titles: {type: object, additionalProperties: {type: string}}
        editions: {type: array, items: {$ref: '#/components/schemas/EditionSummary'}}
        cover: {$ref: '#/components/schemas/MediaSummary'}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
...
    SeriesMember:
      type: object
      required: [series_id, game_id, relation_type, sort_order, game]
      properties:
        series_id: {type: string, format: uuid}
        game_id: {type: string, format: uuid}
        relation_type: {type: string, enum: [mainline, port, remake, spinoff, collection, other]}
        sort_order: {type: integer}
        game: {$ref: '#/components/schemas/Game'}
    SeriesMemberMutation:
      type: object
      additionalProperties: false
      required: [game_id]
      properties:
        game_id: {type: string, format: uuid}
        relation_type: {type: string, enum: [mainline, port, remake, spinoff, collection, other], default: mainline}
        sort_order: {type: integer, minimum: 0, default: 0}
    SeriesMutation:
      type: object
      additionalProperties: false
      required: [default_title]
      properties:
        id: {type: string, format: uuid}
        default_title: {type: string, minLength: 1}
        description: {type: string}
        titles: {type: object, additionalProperties: {type: string}}
        members:
          type: array
          description: When present, replaces the complete member set in the same transaction as metadata.
          items: {$ref: '#/components/schemas/SeriesMemberMutation'}
    Series:
      type: object
      required: [id, default_title, display_title, titles, members, created_at, updated_at]
      properties:
        id: {type: string, format: uuid}
        default_title: {type: string}
        display_title: {type: string}
        description: {type: string}
        titles: {type: object, additionalProperties: {type: string}}
        members: {type: array, items: {$ref: '#/components/schemas/SeriesMember'}}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    SeriesMemberSummary:
      type: object
      additionalProperties: false
      required: [series_id, game_id, relation_type, sort_order, game]
      properties:
        series_id: {type: string, format: uuid}
        game_id: {type: string, format: uuid}
        relation_type: {type: string, enum: [mainline, port, remake, spinoff, collection, other]}
        sort_order: {type: integer}
        game: {$ref: '#/components/schemas/GameSummary'}
    SeriesSummary:
      type: object
      additionalProperties: false
      required: [id, default_title, display_title, titles, members, created_at, updated_at]
      properties:
        id: {type: string, format: uuid}
        default_title: {type: string}
        display_title: {type: string}
        description: {type: string}
        titles: {type: object, additionalProperties: {type: string}}
        members: {type: array, items: {$ref: '#/components/schemas/SeriesMemberSummary'}}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
...
    SeriesCollection:
      allOf:
        - {$ref: '#/components/schemas/PaginatedBase'}
        - {type: object, properties: {data: {type: array, items: {$ref: '#/components/schemas/Series'}}}}
    SeriesSummaryCollection:
      allOf:
        - {$ref: '#/components/schemas/PaginatedBase'}
        - {type: object, properties: {data: {type: array, items: {$ref: '#/components/schemas/SeriesSummary'}}}}
...
    DeviceCollection:
      allOf:
        - {$ref: '#/components/schemas/PaginatedBase'}
        - {type: object, properties: {data: {type: array, items: {$ref: '#/components/schemas/Device'}}}}
    SaveRevisionCollection:
      allOf:
        - {$ref: '#/components/schemas/PaginatedBase'}
        - {type: object, properties: {data: {type: array, items: {$ref: '#/components/schemas/SaveRevision'}}}}
...
    PaginatedBase:
      type: object
      required: [data, pagination]
      properties:
        data: {type: array, items: {type: object}}
        pagination: {$ref: '#/components/schemas/Pagination'}